+++ clang-tools-extra/clang-tidy/cppcoreguidelines/CppCoreGuidelinesTidyModule.cpp
@@ -18,6 +18,7 @@
 #include "AvoidGotoCheck.h"
 #include "AvoidInitDefaultConstructorsCheck.h"
 #include "AvoidNonConstGlobalVariablesCheck.h"
+#include "DeclareLoopVariableInTheInitializerCheck.h"
 #include "InitVariablesCheck.h"
 #include "InterfacesGlobalInitCheck.h"
 #include "MacroUsageCheck.h"
@@ -58,6 +59,8 @@ public:
         "cppcoreguidelines-avoid-magic-numbers");
     CheckFactories.registerCheck<AvoidNonConstGlobalVariablesCheck>(
         "cppcoreguidelines-avoid-non-const-global-variables");
+    CheckFactories.registerCheck<DeclareLoopVariableInTheInitializerCheck>(
+        "cppcoreguidelines-declare-loop-variable-in-the-initializer");
     CheckFactories.registerCheck<modernize::UseOverrideCheck>(
         "cppcoreguidelines-explicit-virtual-functions");
     CheckFactories.registerCheck<InitVariablesCheck>(